improvements and polishing in many-to-many relation

+++ src/Command/Database/DeleteCommand.php
@@ -11,6 +11,7 @@ namespace Spiral\ORM\Command\Database;
 use Spiral\Database\DatabaseInterface;
 use Spiral\ORM\Command\Database\Traits\WhereTrait;
 use Spiral\ORM\Command\ScopedInterface;
+use Spiral\ORM\Exception\CommandException;
 
 class DeleteCommand extends DatabaseCommand implements ScopedInterface
 {
@@ -21,7 +22,7 @@ class DeleteCommand extends DatabaseCommand implements ScopedInterface
      * @param string            $table
      * @param array             $where
      */
-    public function __construct(DatabaseInterface $db, string $table, array $where)
+    public function __construct(DatabaseInterface $db, string $table, array $where = [])
     {
         parent::__construct($db, $table);
         $this->where = $where;
@@ -32,8 +33,11 @@ class DeleteCommand extends DatabaseCommand implements ScopedInterface
      */
     public function execute()
     {
-        $this->db->delete($this->table, $this->where)->run();
+        if (empty($this->where)) {
+            throw new CommandException("Unable to execute delete command without a scope");
+        }
 
+        $this->db->delete($this->table, $this->where)->run();
         parent::execute();
     }
 }
